[#73]/fix rename mainFolders -> subFolders
